+++ proto/gen/rill/runtime/v1/queries.pb.validate.go
@@ -6882,35 +6882,6 @@ func (m *ColumnTimeSeriesRequest) validate(all bool) error {
 		}
 	}
 
-	if all {
-		switch v := interface{}(m.GetFilters()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, ColumnTimeSeriesRequestValidationError{
-					field:  "Filters",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, ColumnTimeSeriesRequestValidationError{
-					field:  "Filters",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetFilters()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return ColumnTimeSeriesRequestValidationError{
-				field:  "Filters",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
 	// no validation rules for Pixels
 
 	// no validation rules for SampleSize
