+++ proto/rill/runtime/v1/queries.proto
@@ -529,7 +529,6 @@ message ColumnTimeSeriesRequest {
   repeated BasicMeasure measures = 3;
   string timestamp_column_name = 4;
   TimeSeriesTimeRange time_range = 5;
-  MetricsViewFilter filters = 6;
   int32 pixels = 7;
   int32 sample_size = 8;
   int32 priority = 9;
