+++ runtime/queries/metricsview.go
@@ -123,12 +123,12 @@ func structTypeToMetricsViewColumn(v *runtimev1.StructType) []*runtimev1.Metrics
 // buildFilterClauseForMetricsViewFilter builds a SQL string of conditions joined with AND.
 // Unless the result is empty, it is prefixed with "AND".
 // I.e. it has the format "AND (...) AND (...) ...".
-func buildFilterClauseForMetricsViewFilter(filter *runtimev1.MetricsViewFilter, dimNameToColMap map[string]string, dialect drivers.Dialect) (string, []any, error) {
+func buildFilterClauseForMetricsViewFilter(mv *runtimev1.MetricsView, filter *runtimev1.MetricsViewFilter, dialect drivers.Dialect) (string, []any, error) {
 	var clauses []string
 	var args []any
 
 	if filter != nil && filter.Include != nil {
-		clause, clauseArgs, err := buildFilterClauseForConditions(filter.Include, dimNameToColMap, false, dialect)
+		clause, clauseArgs, err := buildFilterClauseForConditions(mv, filter.Include, false, dialect)
 		if err != nil {
 			return "", nil, err
 		}
@@ -137,7 +137,7 @@ func buildFilterClauseForMetricsViewFilter(filter *runtimev1.MetricsViewFilter, 
 	}
 
 	if filter != nil && filter.Exclude != nil {
-		clause, clauseArgs, err := buildFilterClauseForConditions(filter.Exclude, dimNameToColMap, true, dialect)
+		clause, clauseArgs, err := buildFilterClauseForConditions(mv, filter.Exclude, true, dialect)
 		if err != nil {
 			return "", nil, err
 		}
@@ -149,17 +149,12 @@ func buildFilterClauseForMetricsViewFilter(filter *runtimev1.MetricsViewFilter, 
 }
 
 // buildFilterClauseForConditions returns a string with the format "AND (...) AND (...) ..."
-func buildFilterClauseForConditions(
-	conds []*runtimev1.MetricsViewFilter_Cond,
-	dimNameToColMap map[string]string,
-	exclude bool,
-	dialect drivers.Dialect,
-) (string, []any, error) {
+func buildFilterClauseForConditions(mv *runtimev1.MetricsView, conds []*runtimev1.MetricsViewFilter_Cond, exclude bool, dialect drivers.Dialect) (string, []any, error) {
 	var clauses []string
 	var args []any
 
 	for _, cond := range conds {
-		condClause, condArgs, err := buildFilterClauseForCondition(cond, dimNameToColMap, exclude, dialect)
+		condClause, condArgs, err := buildFilterClauseForCondition(mv, cond, exclude, dialect)
 		if err != nil {
 			return "", nil, err
 		}
@@ -174,19 +169,15 @@ func buildFilterClauseForConditions(
 }
 
 // buildFilterClauseForCondition returns a string with the format "AND (...)"
-func buildFilterClauseForCondition(
-	cond *runtimev1.MetricsViewFilter_Cond,
-	dimNameToColMap map[string]string,
-	exclude bool,
-	dialect drivers.Dialect,
-) (string, []any, error) {
+func buildFilterClauseForCondition(mv *runtimev1.MetricsView, cond *runtimev1.MetricsViewFilter_Cond, exclude bool, dialect drivers.Dialect) (string, []any, error) {
 	var clauses []string
 	var args []any
 
-	name, ok := dimNameToColMap[cond.Name]
-	if !ok {
-		return "", nil, fmt.Errorf("dimension %s not found", cond.Name)
+	name, err := metricsViewDimensionToSafeColumn(mv, cond.Name)
+	if err != nil {
+		return "", nil, err
 	}
+
 	notKeyword := ""
 	if exclude {
 		notKeyword = "NOT"
@@ -286,20 +277,4 @@ func metricsViewDimensionToSafeColumn(mv *runtimev1.MetricsView, dimName string)
 		}
 	}
 	return "", fmt.Errorf("dimension %s not found", dimName)
-}
-
-// metricsViewDimensionNameMap creates a map of dimension name to safe name of column
-func metricsViewDimensionNameMap(mv *runtimev1.MetricsView) map[string]string {
-	dimNameToColMap := make(map[string]string)
-	for _, dimension := range mv.Dimensions {
-		if dimension.Column != "" {
-			dimNameToColMap[dimension.Name] = safeName(dimension.Column)
-		} else {
-			// backwards compatibility for older projects that have not run reconcile on this dashboard
-			// in that case `column` will not be present
-			dimNameToColMap[dimension.Name] = safeName(dimension.Name)
-		}
-	}
-
-	return dimNameToColMap
 }
