+++ runtime/server/queries_timeseries_test.go
@@ -34,41 +34,6 @@ func TestServer_Timeseries_EmptyModel(t *testing.T) {
 	require.Nil(t, results)
 }
 
-func TestServer_Timeseries(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "max(clicks)",
-				SqlName:    "max",
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Start:    parseTimeToProtoTimeStamps(t, "2019-01-01T00:00:00Z"),
-			End:      parseTimeToProtoTimeStamps(t, "2019-12-02T00:00:00Z"),
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "device",
-					In:   []*structpb.Value{structpb.NewStringValue("android"), structpb.NewStringValue("iphone")},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 1.0, results[0].Records.Fields["max"].GetNumberValue())
-}
-
 func TestServer_Timeseries_Spark_NoParams(t *testing.T) {
 	t.Parallel()
 	server, instanceID := getTimeseriesTestServer(t)
@@ -126,338 +91,6 @@ func isNull(v *structpb.Value) bool {
 	return ok
 }
 
-func Ignore_TestServer_Timeseries_exclude_notnull(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "count(*)",
-				SqlName:    "count",
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Exclude: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "latitude",
-					In:   []*structpb.Value{structpb.NewNumberValue(25)},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 1.0, results[0].Records.Fields["count"])
-}
-
-func Ignore_TestServer_Timeseries_exclude_all(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "count(*)",
-				SqlName:    "count",
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Exclude: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "latitude",
-					In:   []*structpb.Value{structpb.NewNumberValue(25), structpb.NewNullValue()},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 0.0, results[0].Records.Fields["count"])
-}
-
-func TestServer_Timeseries_exclude_notnull_string(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "count(*)",
-				SqlName:    "count",
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Exclude: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "country",
-					In:   []*structpb.Value{structpb.NewStringValue("Canada")},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 1.0, results[0].Records.Fields["count"].GetNumberValue())
-}
-
-func TestServer_Timeseries_exclude_all_string(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "sum(imps)",
-				SqlName:    "Total impressions",
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Exclude: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "country",
-					In:   []*structpb.Value{structpb.NewStringValue("Canada"), structpb.NewNullValue()},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 0.0, results[0].Records.Fields["Total impressions"].GetNumberValue())
-}
-
-func TestServer_Timeseries_exclude_notnull_like(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "count(*)",
-				SqlName:    "count",
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Exclude: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "device",
-					Like: []string{"iphone"},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 1.0, results[0].Records.Fields["count"].GetNumberValue())
-}
-
-func TestServer_Timeseries_exclude_like_all(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "sum(imps)",
-				SqlName:    "Total impressions",
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Exclude: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "country",
-					In:   []*structpb.Value{structpb.NewNullValue()},
-					Like: []string{"Canada"},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 0.0, results[0].Records.Fields["Total impressions"].GetNumberValue())
-}
-
-func TestServer_Timeseries_numeric_dim(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "count(*)",
-				SqlName:    "count",
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "latitude",
-					In:   []*structpb.Value{structpb.NewNumberValue(25)},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 1.0, results[0].Records.Fields["count"].GetNumberValue())
-}
-
-func TestServer_Timeseries_numeric_dim_2values(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "count(*)",
-				SqlName:    "count",
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "latitude",
-					In:   []*structpb.Value{structpb.NewNumberValue(25), structpb.NewNumberValue(35)},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 1.0, results[0].Records.Fields["count"].GetNumberValue())
-}
-
-func TestServer_Timeseries_numeric_dim_and_null(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "count(*)",
-				SqlName:    "count",
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "latitude",
-					In:   []*structpb.Value{structpb.NewNumberValue(25), structpb.NewNullValue()},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 2.0, results[0].Records.Fields["count"].GetNumberValue())
-}
-
-func TestServer_Timeseries_Empty_TimeRange(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "max(clicks)",
-				SqlName:    "max",
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange:           new(runtimev1.TimeSeriesTimeRange),
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "device",
-					In:   []*structpb.Value{structpb.NewStringValue("android"), structpb.NewStringValue("iphone")},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	for i, v := range response.GetRollup().Results {
-		fmt.Printf("i: %d, ts: %v\n", i, v.Ts.AsTime())
-	}
-	require.Equal(t, 25, len(results))
-	require.Equal(t, 1.0, results[0].Records.Fields["max"].GetNumberValue())
-}
-
 func TestServer_Timeseries_Empty_Filter(t *testing.T) {
 	t.Parallel()
 	server, instanceID := getTimeseriesTestServer(t)
@@ -478,7 +111,6 @@ func TestServer_Timeseries_Empty_Filter(t *testing.T) {
 			End:      parseTimeToProtoTimeStamps(t, "2019-12-01T00:00:00Z"),
 			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
 		},
-		Filters: new(runtimev1.MetricsViewFilter),
 	})
 
 	require.NoError(t, err)
@@ -506,7 +138,6 @@ func TestServer_Timeseries_TimeEnd_exclusive(t *testing.T) {
 			End:      parseTimeToProtoTimeStamps(t, "2019-01-02T00:00:00Z"),
 			Interval: runtimev1.TimeGrain_TIME_GRAIN_DAY,
 		},
-		Filters: new(runtimev1.MetricsViewFilter),
 	})
 
 	require.NoError(t, err)
@@ -529,7 +160,6 @@ func TestServer_Timeseries_No_Measures(t *testing.T) {
 			End:      parseTimeToProtoTimeStamps(t, "2019-12-01T00:00:00Z"),
 			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
 		},
-		Filters: new(runtimev1.MetricsViewFilter),
 	})
 
 	require.NoError(t, err)
@@ -551,276 +181,12 @@ func TestServer_Timeseries_Nil_Measures(t *testing.T) {
 			End:      parseTimeToProtoTimeStamps(t, "2019-12-01T00:00:00Z"),
 			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
 		},
-		Filters: new(runtimev1.MetricsViewFilter),
 	})
 
 	require.NoError(t, err)
 	results := response.GetRollup().Results
 	require.Equal(t, 1, len(results))
 	require.Equal(t, 2.0, results[0].Records.Fields["count"].GetNumberValue())
-}
-
-func TestServer_Timeseries_2measures(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	mx := "max"
-	sm := "sum"
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "max(clicks)",
-				SqlName:    mx,
-			},
-			{
-				Expression: "sum(clicks)",
-				SqlName:    sm,
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Start:    parseTimeToProtoTimeStamps(t, "2019-01-01T00:00:00Z"),
-			End:      parseTimeToProtoTimeStamps(t, "2019-12-01T00:00:00Z"),
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "device",
-					In:   []*structpb.Value{structpb.NewStringValue("android"), structpb.NewStringValue("iphone")},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 1.0, results[0].Records.Fields["max"].GetNumberValue())
-	require.Equal(t, 2.0, results[0].Records.Fields["sum"].GetNumberValue())
-}
-
-func TestServer_Timeseries_1dim(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	sm := "sum"
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "sum(clicks)",
-				SqlName:    sm,
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Start:    parseTimeToProtoTimeStamps(t, "2019-01-01T00:00:00Z"),
-			End:      parseTimeToProtoTimeStamps(t, "2019-12-01T00:00:00Z"),
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "device",
-					In:   []*structpb.Value{structpb.NewStringValue("android")},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 1.0, results[0].Records.Fields["sum"].GetNumberValue())
-}
-
-func TestServer_Timeseries_1dim_null(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "sum(clicks)",
-				SqlName:    "sum",
-			},
-		},
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		TimestampColumnName: "time",
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "publisher",
-					In:   []*structpb.Value{structpb.NewNullValue()},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 1.0, results[0].Records.Fields["sum"].GetNumberValue())
-}
-
-func TestServer_Timeseries_1dim_null_and_in(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "sum(clicks)",
-				SqlName:    "sum",
-			},
-		},
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		TimestampColumnName: "time",
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "publisher",
-					In: []*structpb.Value{
-						structpb.NewNullValue(),
-						structpb.NewStringValue("Google"),
-					},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 2.0, results[0].Records.Fields["sum"].GetNumberValue())
-}
-
-func TestServer_Timeseries_1dim_null_and_in_and_like(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "sum(clicks)",
-				SqlName:    "sum",
-			},
-		},
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		TimestampColumnName: "time",
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "publisher",
-					In: []*structpb.Value{
-						structpb.NewNullValue(),
-						structpb.NewStringValue("Google"),
-					},
-					Like: []string{
-						"Goo%",
-					},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 2.0, results[0].Records.Fields["sum"].GetNumberValue())
-}
-
-func TestServer_Timeseries_1dim_2like(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "sum(clicks)",
-				SqlName:    "sum",
-			},
-		},
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		TimestampColumnName: "time",
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "domain",
-					Like: []string{
-						"g%",
-						"msn%",
-					},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 2.0, results[0].Records.Fields["sum"].GetNumberValue())
-}
-
-func TestServer_Timeseries_2dim_include_and_exclude(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "sum(clicks)",
-				SqlName:    "sum",
-			},
-		},
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_YEAR,
-		},
-		TimestampColumnName: "time",
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "publisher",
-					In: []*structpb.Value{
-						structpb.NewStringValue("Google"),
-					},
-				},
-				{
-					Name: "domain",
-					In: []*structpb.Value{
-						structpb.NewStringValue("msn.com"),
-					},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 1, len(results))
-	require.Equal(t, 0.0, results[0].Records.Fields["sum"].GetNumberValue())
 }
 
 func TestServer_Timeseries_no_measures(t *testing.T) {
@@ -835,77 +201,6 @@ func TestServer_Timeseries_no_measures(t *testing.T) {
 			Start:    parseTimeToProtoTimeStamps(t, "2019-01-01T00:00:00Z"),
 			End:      parseTimeToProtoTimeStamps(t, "2019-01-03T00:00:00Z"),
 			Interval: runtimev1.TimeGrain_TIME_GRAIN_DAY,
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 2, len(results))
-	require.Equal(t, 1.0, results[0].Records.Fields["count"].GetNumberValue())
-}
-
-func TestServer_Timeseries_1day(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	mx := "max"
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "max(clicks)",
-				SqlName:    mx,
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Start:    parseTimeToProtoTimeStamps(t, "2019-01-01T00:00:00Z"),
-			End:      parseTimeToProtoTimeStamps(t, "2019-01-03T00:00:00Z"),
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_DAY,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "device",
-					In:   []*structpb.Value{structpb.NewStringValue("android"), structpb.NewStringValue("iphone")},
-				},
-			},
-		},
-	})
-
-	require.NoError(t, err)
-	results := response.GetRollup().Results
-	require.Equal(t, 2, len(results))
-}
-
-func TestServer_Timeseries_1day_Count(t *testing.T) {
-	t.Parallel()
-	server, instanceID := getTimeseriesTestServer(t)
-
-	cnt := "count"
-	response, err := server.ColumnTimeSeries(testCtx(), &runtimev1.ColumnTimeSeriesRequest{
-		InstanceId: instanceID,
-		TableName:  "timeseries",
-		Measures: []*runtimev1.ColumnTimeSeriesRequest_BasicMeasure{
-			{
-				Expression: "count(*)",
-				SqlName:    cnt,
-			},
-		},
-		TimestampColumnName: "time",
-		TimeRange: &runtimev1.TimeSeriesTimeRange{
-			Start:    parseTimeToProtoTimeStamps(t, "2019-01-01T00:00:00Z"),
-			End:      parseTimeToProtoTimeStamps(t, "2019-01-03T00:00:00Z"),
-			Interval: runtimev1.TimeGrain_TIME_GRAIN_DAY,
-		},
-		Filters: &runtimev1.MetricsViewFilter{
-			Include: []*runtimev1.MetricsViewFilter_Cond{
-				{
-					Name: "device",
-					In:   []*structpb.Value{structpb.NewStringValue("android"), structpb.NewStringValue("iphone")},
-				},
-			},
 		},
 	})
